Use {{region_code}} in address_format
The {{region}} placeholder was used ambiguously in address_format.

According to our discussion in /hexorx/countries/pull/221 we agreed on
using different placeholders, {{region}} or {{region_code}}, depending
on whether the full region name (like 'Bayern' for Germany) or the
abbrevation (like 'NY' in US address format) should be used.

This commit replaces any occurence of {{region}} by {{region_code}}
where {{region}} was not used on a separate line in the YAML files.