Fix latitude/longitude details for New York, US.
When Google's geocoder is passed "New York, United States", it actually
returns the geocoding information for New York City, and not the actual
state. So the previously stored details were for the extent of the city,
rather than the state. This fixes the initial geocoding logic by making
a special case for New York to geocode "New York State". The stored YAML
details have been updated with the new results for the state.

I've seen Google's geocoder change how it geocodes the string "New York"
over the years, but appending "State" to the end has been a reliable way
to force it to geocode the state (at least in my experience for the past
couple years). So I think this approach should be safe in case anyone
want to do a full re-geocoding sometime in the future.