Run rake scripts to update and clean CLDR subdivision data
This leaves out changes to `ccp` translations. Which make the commit difficult to audit.

The changes to the `ccp` translations seem to stem primarily from calls made
with the `yaml` library. The library appears to encode all Chakma glyphs as an
escaped sequence of unicode rather than encoding the text using UTF. So a lot of
these changes are not tangible changes but rather just a change in encoding.